Move AuthRepository.kt into repository package.
