+++ app/src/main/java/com/appunite/loudius/di/GithubModule.kt
@@ -1,9 +1,9 @@
 package com.appunite.loudius.di
 
 import android.content.Context
-import com.appunite.loudius.domain.AuthRepository
-import com.appunite.loudius.domain.AuthRepositoryImpl
 import com.appunite.loudius.domain.UserLocalDataSource
+import com.appunite.loudius.domain.repository.AuthRepository
+import com.appunite.loudius.domain.repository.AuthRepositoryImpl
 import com.appunite.loudius.network.datasource.AuthDataSource
 import com.appunite.loudius.network.datasource.AuthNetworkDataSource
 import com.appunite.loudius.network.services.AuthService
@@ -14,8 +14,8 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
 import javax.inject.Singleton
+import retrofit2.Retrofit
 
 @InstallIn(SingletonComponent::class)
 @Module
