+++ app/src/test/java/com/appunite/loudius/domain/AuthRepositoryImplTest.kt
@@ -1,5 +1,6 @@
 package com.appunite.loudius.domain
 
+import com.appunite.loudius.domain.repository.AuthRepositoryImpl
 import com.appunite.loudius.network.datasource.AuthDataSource
 import io.mockk.coEvery
 import io.mockk.coVerify
