+++ app/src/test/java/com/appunite/loudius/network/NetworkTestDoubles.kt
@@ -1,18 +1,18 @@
 package com.appunite.loudius.network
 
-import com.appunite.loudius.domain.AuthRepository
+import com.appunite.loudius.domain.repository.AuthRepository
 import com.appunite.loudius.fakes.FakeAuthRepository
 import com.appunite.loudius.network.utils.AuthInterceptor
 import com.appunite.loudius.network.utils.LocalDateTimeDeserializer
 import com.google.gson.FieldNamingPolicy
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import java.time.LocalDateTime
+import java.util.concurrent.TimeUnit
 import okhttp3.OkHttpClient
 import okhttp3.mockwebserver.MockWebServer
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import java.time.LocalDateTime
-import java.util.concurrent.TimeUnit
 
 fun testOkHttpClient(authRepository: AuthRepository = FakeAuthRepository()) =
     OkHttpClient.Builder()
